# generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ExerciseLogResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"log": {
						"$ref": "#/components/schemas/ExerciseLog"
					}
				},
				"required": [
					"success",
					"log"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseLogsResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"logs": {
						"items": {
							"$ref": "#/components/schemas/ExerciseLog"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"logs"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BlockLogResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"log": {
						"$ref": "#/components/schemas/BlockLog"
					}
				},
				"required": [
					"success",
					"log"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "mastersfit-server",
		"version": "1.0.0",
		"description": "RESTful API server for Masters Fit fitness tracking application",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/logs/exercise": {
			"post": {
				"operationId": "CreateExerciseLog",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create exercise log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Exercise log data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Exercise log data"
							}
						}
					}
				}
			}
		},
		"/logs/exercise/{exerciseLogId}": {
			"put": {
				"operationId": "UpdateExerciseLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update exercise log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise log ID",
						"in": "path",
						"name": "exerciseLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Update data"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetExerciseLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise log by ID",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise log ID",
						"in": "path",
						"name": "exerciseLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/exercise/plan-day-exercise/{planDayExerciseId}": {
			"get": {
				"operationId": "GetExerciseLogsForPlanDayExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise logs for a plan day exercise",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day exercise ID",
						"in": "path",
						"name": "planDayExerciseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/exercise/block/{workoutBlockId}": {
			"get": {
				"operationId": "GetExerciseLogsForWorkoutBlock",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise logs for a workout block",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Workout block ID",
						"in": "path",
						"name": "workoutBlockId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/exercise/plan-day/{planDayId}": {
			"get": {
				"operationId": "GetExerciseLogsForPlanDay",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise logs for a plan day (legacy endpoint)",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day ID",
						"in": "path",
						"name": "planDayId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/block": {
			"post": {
				"operationId": "CreateBlockLog",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create block log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Block log data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Block log data"
							}
						}
					}
				}
			}
		},
		"/logs/block/{blockLogId}": {
			"put": {
				"operationId": "UpdateBlockLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update block log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Block log ID",
						"in": "path",
						"name": "blockLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Update data"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBlockLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get block log by ID",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Block log ID",
						"in": "path",
						"name": "blockLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/block/workout-block/{workoutBlockId}": {
			"get": {
				"operationId": "GetBlockLogsForWorkoutBlock",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get block logs for a workout block",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Workout block ID",
						"in": "path",
						"name": "workoutBlockId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/block/workout-block/{workoutBlockId}/latest": {
			"get": {
				"operationId": "GetLatestBlockLogForWorkoutBlock",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get latest block log for a workout block",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Workout block ID",
						"in": "path",
						"name": "workoutBlockId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/block/plan-day/{planDayId}": {
			"get": {
				"operationId": "GetBlockLogsForPlanDay",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlockLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get block logs for a plan day",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day ID",
						"in": "path",
						"name": "planDayId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/plan-day": {
			"post": {
				"operationId": "CreatePlanDayLog",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create plan day log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Plan day log data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Plan day log data"
							}
						}
					}
				}
			}
		},
		"/logs/plan-day/{planDayLogId}": {
			"put": {
				"operationId": "UpdatePlanDayLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update plan day log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day log ID",
						"in": "path",
						"name": "planDayLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Update data"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPlanDayLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get plan day log by ID",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day log ID",
						"in": "path",
						"name": "planDayLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/plan-day/plan-day/{planDayId}": {
			"get": {
				"operationId": "GetPlanDayLogsForPlanDay",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get plan day logs for a plan day",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day ID",
						"in": "path",
						"name": "planDayId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/plan-day/plan-day/{planDayId}/latest": {
			"get": {
				"operationId": "GetLatestPlanDayLogForPlanDay",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get latest plan day log for a plan day",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day ID",
						"in": "path",
						"name": "planDayId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/logs/plan-day/workout/{workoutId}": {
			"get": {
				"operationId": "GetPlanDayLogsForWorkout",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlanDayLogsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get plan day logs for a workout",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Workout ID",
						"in": "path",
						"name": "workoutId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}